Migrate Error Handling Middleware
