+++ api/src/controllers/ErrorController.ts
@@ -1,27 +1,4 @@
-import { Request, Response, NextFunction } from "express";
-
-type GenericError = {
-  statusCode: number;
-  message: string;
-};
-
 export default class ErrorController {
-  static errorHandler(
-    err: GenericError,
-    req: Request,
-    res: Response,
-    next: NextFunction
-  ) {
-    const statusCode = err.statusCode || 500;
-    const message = err.message || "Internal Server Error";
-
-    res.status(statusCode).json({
-      error: {
-        message: message,
-      },
-    });
-  }
-
   public static InternalServerError(message: string = "Internal Server Error") {
     throw new HandleError(message, 500);
   }
@@ -40,7 +17,7 @@ export default class ErrorController {
 }
 
 class HandleError extends Error {
-  private readonly statusCode: number;
+  statusCode: number;
   constructor(message: string, statusCode: number) {
     super(message);
     this.statusCode = statusCode;
